move template shared logic
