+++ src/utils/resources/template/hooks/index.ts
@@ -1,1 +1,3 @@
-export {};
+export * from './useProcessedTemplate';
+export * from './useVmTemplates';
+export * from './useVmTemplateSource';
