+++ src/utils/resources/template/hooks/useProcessedTemplate.ts
