+++ src/utils/resources/template/hooks/useVmTemplateSource/index.ts
@@ -1,0 +1,1 @@
+export * from './useVmTemplateSource';
