+++ src/utils/resources/template/utils/annotations.ts
