+++ src/utils/resources/template/utils/index.ts
@@ -1,1 +1,4 @@
-export {};
+export * from './annotations';
+export * from './constants';
+export * from './flavor';
+export * from './selectors';
